% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_download.R
\name{download_files}
\alias{download_files}
\title{Download specified protocol files from the bundestag website}
\usage{
download_files(links, directory = "./protokolle")
}
\arguments{
\item{links}{A vector of characters containing links to the files to download}

\item{directory}{Directory where the files should be stored}
}
\value{
No return values but the downloaded protocols are stored in the specified directory
}
\description{
Iterates over a vector of links to files and downloads them into the target directory
}
\examples{
parse_protocols(links = c("https://www.bundestag.de/resource/blob/704800/a2bb9229c2e19bbd633d30c747b41c62/19171-data.xml","https://www.bundestag.de/resource/blob/704394/b1d0626827462b96db4e2b7821b1c0cb/19170-data.xml"))

}
