% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_download.R
\name{download_protocols}
\alias{download_protocols}
\title{Wraps get_protocol_links and download_files to get all protocol files in a single call}
\usage{
download_protocols(
  base_url = "https://www.bundestag.de",
  registry_url = "/ajax/filterlist/de/services/opendata/543410-543410",
  directory = "./protokolle"
)
}
\arguments{
\item{base_url}{The URL of the bundestag.de-domain this should continue to be "https://www.bundestag.de}

\item{registry_url}{This is the path to the service the "https://www.bundestag.de/protokolle"-frontend hits, to get the links to the protocol files.
The current default value is equivalent to the service path for the 19th period. This may change in the future.
It is necessary to use the devtools of the browser to get this information.}
}
\value{
No return values but the downloaded protocols are stored in the specified directory
}
\description{
This function gets all download links and downloads the linked files into the specified folder
}
\examples{
download_protocols()

}
