% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_download.R
\name{get_protocol_links}
\alias{get_protocol_links}
\title{Scrape the download-links for the plenary protocols from the bundestag website}
\usage{
get_protocol_links(
  base_url = "https://www.bundestag.de",
  registry_url = "/ajax/filterlist/de/services/opendata/543410-543410"
)
}
\arguments{
\item{base_url}{The URL of the bundestag.de-domain this should continue to be "https://www.bundestag.de}

\item{registry_url}{This is the path to the service the "https://www.bundestag.de/protokolle"-frontend hits, to get the links to the protocol files.
The current default value is equivalent to the service path for the 19th period. This may change in the future.
It is necessary to use the devtools of the browser to get this information.}

\item{directory}{Directory where the files should be stored}
}
\value{
A vector of complete links to the protocol files.
}
\description{
This function queries a file-registry-enpoint of the bundestag.de website repeadetly, to scrape the direct download links.
}
\examples{
get_protocol_links()

}
