% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph_list.R
\name{paragraph_list}
\alias{paragraph_list}
\title{Parse a plenary protocol from xml for paragraphs into tibbles
Uses the xml structure of a plenary protocol to create two tibbles for further analysis}
\usage{
paragraph_list(protocol)
}
\arguments{
\item{protocol}{a XML nodeset as returned by xml2 containing a plenary protocoll}
}
\value{
A tibble containing all officially spoken paragraphs during a plenary speech.
Including the main speach, moderatory interjections and questions by parliament members
}
\description{
Parse a plenary protocol from xml for paragraphs into tibbles
Uses the xml structure of a plenary protocol to create two tibbles for further analysis
}
\examples{
paragraph_list(read_xml("./protokolle/19007-data.xml"))
}
