% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protocol.R
\name{parse_protocol}
\alias{parse_protocol}
\title{Parse a plenary protocol from xml format to tibbles}
\usage{
parse_protocol(path, check_schema = TRUE)
}
\arguments{
\item{path}{A string containing the path to the xml file you want to parse}

\item{check_schema}{A logical value indicating whether you want to check compatibility of the xml schema used in your xml file.}
}
\value{
Three tibbles in a named list:

"speakers": A tibble of all speaking politicians containing speaker id, name, party and similar information.

"paragraphs": A tibble of all paragraphs in speeches, containing speaker id, speech id and content of the paragraph.

"comments": A tibble of all comments given during speeches and about reactions to speeches, containing speech id and comment id as well as content of the comment.
}
\description{
Uses the xml structure of a plenary protocol to create three tibbles for further data analysis.
}
