% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protokoll.R
\name{parse_protocols}
\alias{parse_protocols}
\title{Parse all plenary protocols from xml format to tibbles}
\usage{
parse_protocols(path = "protokolle", start = NULL, end = NULL)
}
\arguments{
\item{path}{A string containing the path to the xml files you want to parse}

\item{start}{A string of the protocol you want to start with (optional)}

\item{end}{A string of the protocol you want to end with (optional). Must succed start in alphabetical order}
}
\value{
Three tibbles in a named list:

"rednerliste": A tibble of all speaking politicians containing speaker id, name, party and similar information.

"absatzliste": A tibble of all paragraphs in speeches, containing speaker id, speech id and content of the paragraph.

"kommentarliste": A tibble of all comments given during speeches and about reactions to speeches, containing speech id and comment id as well as content of the comment.
}
\description{
Uses the xml structure of a plenary protocols to create three tibbles for further data analysis with the help of parse_protocol().
}
\examples{
parse_protocols(start = "19001-data.xml", end = "19003-data.xml")

}
