% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protokoll.R
\name{parse_protokoll}
\alias{parse_protokoll}
\title{Parse a plenary protocol from xml format to tibbles}
\usage{
parse_protokoll(path)
}
\arguments{
\item{path}{A string containing the path to the xml file you want to parse}
}
\value{
Three tibbles in a named list:

"rednerliste": A tibble of all speaking politicians containing speaker id, name, party and similar information.

"absatzliste": A tibble of all paragraphs in speeches, containing speaker id, speech id and content of the paragraph.

"kommentarliste": A tibble of all comments given during speeches and about reactions to speeches, containing speech id and comment id as well as content of the comment.
}
\description{
Uses the xml structure of a plenary protocol to create three tibbles for further data analysis.
}
\examples{
parse_protokoll("./protokolle/19007-data.xml")

}
