% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_download.R
\name{download_files}
\alias{download_files}
\title{Download specified protocol files from the bundestag website}
\usage{
download_files(links, directory = "./protokolle")
}
\arguments{
\item{links}{A vector of characters containing links to the files to download}

\item{directory}{Directory where the files should be stored}
}
\value{
A vector with all paths to the downloaded files
}
\description{
Iterates over a vector of links to files and downloads them into the target directory
There exist no tests for it, because this function interacts directly with external systems and causing load and expenses for the operator of them.
Testing against them is considered bad practice. I could have mocked the server and it's responses, but this is not worth the effort given the fact that
this function uses unsupported interfaces anyway.
}
\examples{
parse_protocols(links = "https://www.bundestag.de/resource/blob/704800/a2bb9229c2e19bbd633d30c747b41c62/19171-data.xml")

}
