% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_protocol.R
\name{parse_protocols}
\alias{parse_protocols}
\title{Parse all plenary protocols from xml format to tibbles in parallel}
\usage{
parse_protocols(
  path = "protokolle",
  start = NULL,
  end = NULL,
  instance_count = NULL,
  check_schema = TRUE
)
}
\arguments{
\item{path}{A string containing the path to the xml files you want to parse}

\item{start}{Name of the protocol you want to start with (optional)}

\item{end}{Name of the protocol you want to end with (optional). Must succeed start in alphabetical order}

\item{instance_count}{Specifies the number of r instances that will be used to parse the protocols. The default is the machines core count.}

\item{check_schema}{A logical value indicating whether you want to check compatibility of the xml schema used in your xml file.}
}
\value{
Three tibbles in a named list:

"speakers": A tibble of all speaking politicians containing speaker id, name, party and similar information.

"paragraphs": A tibble of all paragraphs in speeches, containing speaker id, speech id and content of the paragraph.

"comments": A tibble of all comments given during speeches and about reactions to speeches, containing speech id and comment id as well as content of the comment.
}
\description{
Uses the xml structure of a plenary protocols to create three tibbles for further data analysis with the help of parse_protocol().
}
\examples{
parse_protocols(start = "19001-data.xml", end = "19003-data.xml")

}
